Add various reductions for parsed perc types
Reduce the types of user defined types and record fields in various
ways:

    - union reduction: necessary to decide which types should be
      treated as a "maybe" and thus omitted if they are undefined
    - reduction of ignored types: used to decide if a field is ignored
      because of an error or an unsupported type definition. The error
      can come from the type tree or in the user-defined types
      referenced by the type.
    - dependency reduction: remove unused record and user type
      definitions and keep only definitions that the exported types
      depend on.

The reduction of unions is "local" to a perc type. It does not need to
follow through to the referenced records/user-defined types. It is
defined directly in the perc_reduce module.

In contrast, the other reductions need to follow those references.
The types thus form a directed graph. The construction and queries on
that graph are made in the perc_digraph module.