Add no_empty_obj filter
This is a special filter because it does not depend directly on the
value it acts upon, but rather on the resulting encoded json
object. Therefore, it cannot purely be implemented as a C function and
it is instead inserted in the generated json-encoding code as a
special case of filtering.