Replace yecc-generated parser by a recursive descent parser
A parser generated by yecc is not composable. Thus it is hard to reuse
parts of the grammar in different places. A recursive descent parser
makes reusability easy.

Having a composable grammar will be useful for parsing perc types from
comments in erlang record definitions, in parallel to the perc file
grammar.